+++ src/alley/wp/alleyvate/load.php
@@ -19,17 +19,20 @@ namespace Alley\WP\Alleyvate;
  */
 function available_features(): array {
 	return [
-		'clean_admin_bar'               => new Features\Clean_Admin_Bar(),
-		'disable_comments'              => new Features\Disable_Comments(),
-		'disable_dashboard_widgets'     => new Features\Disable_Dashboard_Widgets(),
-		'disable_sticky_posts'          => new Features\Disable_Sticky_Posts(),
-		'disable_trackbacks'            => new Features\Disable_Trackbacks(),
-		'disallow_file_edit'            => new Features\Disallow_File_Edit(),
-		'login_nonce'                   => new Features\Login_Nonce(),
-		'redirect_guess_shortcircuit'   => new Features\Redirect_Guess_Shortcircuit(),
-		'site_health'                   => new Features\Site_Health(),
-		'user_enumeration_restrictions' => new Features\User_Enumeration_Restrictions(),
-		'full_page_cache_404'           => new Features\Full_Page_Cache_404(),
+		'clean_admin_bar'                      => new Features\Clean_Admin_Bar(),
+		'disable_attachment_routing'           => new Features\Disable_Attachment_Routing(),
+		'disable_comments'                     => new Features\Disable_Comments(),
+		'disable_custom_fields_meta_box'       => new Features\Disable_Custom_Fields_Meta_Box(),
+		'disable_dashboard_widgets'            => new Features\Disable_Dashboard_Widgets(),
+		'disable_password_change_notification' => new Features\Disable_Password_Change_Notification(),
+		'disable_sticky_posts'                 => new Features\Disable_Sticky_Posts(),
+		'disable_trackbacks'                   => new Features\Disable_Trackbacks(),
+		'disallow_file_edit'                   => new Features\Disallow_File_Edit(),
+		'login_nonce'                          => new Features\Login_Nonce(),
+		'redirect_guess_shortcircuit'          => new Features\Redirect_Guess_Shortcircuit(),
+		'site_health'                          => new Features\Site_Health(),
+		'user_enumeration_restrictions'        => new Features\User_Enumeration_Restrictions(),
+		'full_page_cache_404'                  => new Features\Full_Page_Cache_404(),
 	];
 }